requirements.txt
pandas
numpy
streamlit-drawable-canvas==0.8.0
pydeck
plotly